requirements.txt
Django>=4.2.0
django-crispy-forms>=2.0
python-dateutil>=2.8.2
qrcode>=7.4.2
Pillow>=10.0.0
resend>=0.7.2 
reportlab>=4.0.0
